Move sampleEvents to use-event-data file and return sample data on event api error
